Document django-devserver interaction.
Resolves #7.

It appears the fix for django was in:

https://github.com/django/django/commit/728548e483a5a3486939b0c8e62520296587482e

However, this does not reference the issue I am seeing directly. I suspect
this is because django-devserver and django-boardinghouse are both apps
that do unusual things with the database access stack, and this interaction
causes something to go wrong with the MRO, resulting in a Stack Overflow.

As long as django 1.7 or later is used, this stack overflow is prevented.